+++ packages/cli/templates/Gjs/ui/main.d.ts
@@ -1,7 +1,9 @@
+<% const gio = dep.find('Gio') %>
+
 <%_ if(useNamespace){ _%>
-    import type Gio from '../Gio-2.0.js';
+    import type Gio from '../<%= gio.packageName %>.js';
 <%_ } else { _%>
-    import type * as Gio from '../Gio-2.0.js';
+    import type * as Gio from '../<%= gio.packageName %>.js';
 <%_ } _%>
 
 import { ComponentManager } from './components/__init__.js';