+++ packages/cli/templates/Gjs/ui/popupMenu.d.ts
@@ -1,21 +1,24 @@
 // https://gitlab.gnome.org/GNOME/gnome-shell/-/blob/main/js/ui/popupMenu.js
-<%_ if(useNamespace){ _%>
-    import type Gio from '../Gio-2.0.js';
-    import type St from '../St-1.0.js';
-    import type Shell from '../Shell-0.1.js';
-    // TODO: on Ubuntu 22.04 this is Clutter-10
-    // import type Clutter from '../Clutter-11.js';
-<%_ } else { _%>
-    import type * as Gio from '../Gio-2.0.js';
-    import type * as St from '../St-1.0.js';
-    import type * as Shell from '../Shell-0.1.js';
-    // TODO: on Ubuntu 22.04 this is Clutter-10
-    // import type * as Clutter from '../Clutter-11.js';
-<%_ } _%>
-import * as Signals from '../misc/signals.js';
 
-// TODO: Replace this with the actual ClutterEvent type
-type ClutterEvent = any;
+<% const gio = dep.find('Gio') %>
+<% const st = dep.find('St') %>
+<% const shell = dep.find('Shell') %>
+<% const clutter = dep.find('Clutter') %>
+<% const meta = dep.find('Meta') %>
+
+<%_ if(useNamespace){ _%>
+    import type Gio from '../<%= gio.packageName %>.js';
+    import type St from '../<%= st.packageName %>.js';
+    import type Shell from '../<%= shell.packageName %>.js';
+    import type Clutter from '../<%= clutter.packageName %>.js';
+<%_ } else { _%>
+    import type * as Gio from '../<%= gio.packageName %>.js';
+    import type * as St from '../<%= st.packageName %>.js';
+    import type * as Shell from '../<%= shell.packageName %>.js';
+    import type * as Clutter from '../<%= clutter.packageName %>.js';
+<%_ } _%>
+
+import * as Signals from '../misc/signals.js';
 
 export enum Ornament {
     NONE = 0,
@@ -42,7 +45,7 @@ declare class PopupBaseMenuItem extends St.BoxLayout {
     constructor(params?: PopupBaseMenuItem.ConstructorProperties);
     override _init(...args: any[]): void;
 
-    activate(event: ClutterEvent): void;
+    activate(event: Clutter.Event): void;
 
     // General signal handler methods
     connect(sigName: string, callback: (...args: any[]) => void): number
@@ -51,8 +54,8 @@ declare class PopupBaseMenuItem extends St.BoxLayout {
     disconnect(id: number): void
 
     // Specific signal handler methods
-    connect(sigName: "activate", callback: (($obj: PopupBaseMenuItem, event: ClutterEvent) => void)): number
-    connect_after(sigName: "activate", callback: (($obj: PopupBaseMenuItem, event: ClutterEvent) => void)): number
+    connect(sigName: "activate", callback: (($obj: PopupBaseMenuItem, event: Clutter.Event) => void)): number
+    connect_after(sigName: "activate", callback: (($obj: PopupBaseMenuItem, event: Clutter.Event) => void)): number
 
     syncSensitive(): boolean;
     getSensitive(): boolean;
@@ -106,7 +109,7 @@ export class PopupSwitchMenuItem extends PopupBaseMenuItem {
     override _init(text: string, active: boolean, params?: PopupSwitchMenuItem.ConstructorProperties): void;
 
     setStatus(text: string): void;
-    activate(event: ClutterEvent): void;
+    activate(event: Clutter.Event): void;
     toggle(): void;
     setToggleState(state: boolean): void;
     checkAccessibleState(): void;
@@ -207,7 +210,7 @@ export class PopupSubMenuMenuItem extends PopupBaseMenuItem {
 
     syncSensitive(): boolean;
     setSubmenuShown(open: boolean): void;
-    activate(event: ClutterEvent): void;
+    activate(event: Clutter.Event): void;
 }
 
 export namespace PopupMenuManager {